rx: Redfined eye catcher on AIX + Open XL C
When building the kernel extension on AIX 7.2+ using
the CLANG-based Open XL C 17.1+ compiler, during the
build of src/rx/AIX/rx_knet.c, the following error is
thrown:

/home/build/openafs/src/rx/AIX/rx_knet.c:378:6: error: expected expression
            MGETHDR(m, M_DONTWAIT, MT_DATA);
            ^
../sys/mbuf_macro.h:389:25: note: expanded from macro 'MGETHDR'
        (m)->m_eyecat = EYEC_MBUFA;                                             \
                        ^
../sys/mbuf_base.h:79:61: note: expanded from macro 'EYEC_MBUFA'
                                                            ^

This eye catcher macro in the IBM-supplied header
files is at fault.  Redefine EYEC_MBUFA to include
a space as the final byte.

Change-Id: Ifc6d06b097ccb2ae70f2e7021d12bc595f81249b