tsm41: Use the compiler front-end to link
Currently the Makefile rules for aklog_dynamic_auth,
afs_dynamic_auth, and afs_dynamic_kerbauth use $(LD) to
perform linking.  On AIX with the new clang-based Open
XL C 17.1+ compilers, these targets fail to build due to
missing symbols:

/usr/bin/ld -o afs_dynamic_auth aix_auth.o  aix_ident.o  aix_auth_common.o  aix_ktc_null.o /home/build/openafs/lib/libkauth.a  /home/build/openafs/lib/libprot.a  /home/build/openafs/lib/libubik.a  /home/build/openafs/lib/libauth.a  /home/build/openafs/lib/librxkad.a  /home/build/openafs/lib/libsys.a  /home/build/openafs/lib/libafsrfc3961.a  /home/build/openafs/lib/libafshcrypto_lwp.a  /home/build/openafs/lib/librx.a  /home/build/openafs/lib/liblwp.a  /home/build/openafs/lib/libcmd.a  /home/build/openafs/lib/libafscom_err.a  /home/build/openafs/lib/libaudit.a  /home/build/openafs/lib/util.a  /home/build/openafs/lib/libopr.a   -L/home/build/openafs/lib -lrokenafs -ldl -eafs_initialize -bI:/lib/aio.exp -bI:/lib/netinet.exp -bI:/lib/sockets.exp -bI:/lib/statcmd.exp -lsys -lcsys -lc
ld: 0711-317 ERROR: Undefined symbol: .__udivdi3
ld: 0711-345 Use the -bloadmap or -bnoquiet option to obtain more information.
make: The error code from the last command is 8.

The reference to __udivdi3 is in snprintf.c:

	The following symbols are in error:
 Symbol                    Inpndx  TY CL Source-File(Object-File) OR Import-File{Shared-object}
                              RLD: Address  Section  Rld-type Referencing Symbol
 ----------------------------------------------------------------------------------------------
 .__udivdi3                [11]    ER PR /home/build/openafs/src/external/heimdal/roken/snprintf.c(/home/build/openafs/lib/librokenafs.a[snprintf.o])
                                   00000c7c .text    R_RBR    [47]    <.append_number>

The __udivdi3 symbol is part of the clang runtime and is
found in the following library:

/opt/IBM/openxlC/17.1.0/lib/clang/13.0.0/lib/aix/libclang_rt.builtins-powerpc.a

Using the compiler front-end (ibm-clang) to perform the
link step causes this library to be linked in
automatically.  This works on the old XL C 16.1 and
below compiler versions as well.

Change-Id: I77d544a27cb87ec71665a1add9eb8449fefbf8a1