export: Set libpath to NULL before calling sysconfig
On AIX, kernel extensions are loaded by calling the function
sysconfig and passing it a structure of type cfg_load.  This
structure has two members, path and libpath.

When compiled using Open XL C 17.1+, if the libpath member is
not set to NULL, the following error is displayed:

Could not load program cfgexport:

IBM's sample kernel extension always populates both structure
members.  As we don't require use of the libpath paremeter,
setting it to NULL allows the extension to load successfully.

This does not seem to be necessary when using the older XL C
16.1 compiler and below, however setting it explicitly does
not hurt, either.

Change-Id: I8b7a4b53f8413387a0aa54ed4265161ddb869051