libafs: Remove linker alignment option on AIX
When building the kernel extension on AIX using the new
CLANG-based Open XL C 17.1+ compilers, during the link
stage, the following warning is issued:

ld: 0711-356 WARNING: The .text section may not be aligned properly.

Unfortunately this is correct, and the built extension
does not load into the kernel.

Removing the option from LDFLAGS in MakefileProto.AIX.in
resolves the issue and the linker aligns the .text
section correctly automatically.  This is more similar to
how IBM's demo kernel extension is built, as it does not
specify the -H8 option in it's ld invocation either.

Change-Id: I8aafaed3a620f19953b7730aae7c2b901979d1f3