export: Don't use flex-string format when using ibm-clang
On AIX 7.2+ with the new CLANG-based Open XL C 17.1+
compilers, a kernel panic occurs in the function sym_flex
in src/export/symtab.c on the following line:

    symbol = *sym;

This results in the following kernel panic:

CRASH INFORMATION:
CPU 0 CSA F00000002FF47600 at time of crash, error code for LEDs: 70000000
pvthread+098800 STACK:
[F10009D5B0650F34]symsrch+000108 (F10009D5B06521D0)
[F10009D5B0650CB0]sym_lookup+000024 (000000000017E734, 0000000000000000)
[F10009D5B065087C]import_kfunc+000034 (0000000110000FDB)
[F10009D5B0654A74]kluge_init+000084 ()
[F10009D5B06544F8]afs_config+000088 (0000000000000001, FFFF3FFFFFFFFFF3)
[00014D70].hkey_legacy_gate+00004C ()
[008797C8]IPRA.$config_kmod+000128 (??, ??, ??)
[00879FD4]sysconfig+000194 (??, ??, ??)
[00003984]syscall+000244 ()
[kdb_get_virtual_memory] no real storage @ FFFFFFFFFFFD870
[100000A44]0000000100000A44 ()

While the backtrace indicates the failing function is
symsrch, the final line of symsrch is:

    return found ? sym_flex(found) : 0;

It is actually the call to sym_flex which results in
the panic.

The purpose of the sym_flex function is to change the
format of the n_nptr field in the returned sym_t
structure.  It is unknown why this is necessary as the
upstream code does not seem to reference this field
subsequently.  Therefore, to avoid the kernel panic,
we skip the call entirely when using the Open XL C
compiler and instead just let symsrch return the
pointer directly.

Change-Id: Icf08661cb5ed4cc5879a09a2640cec94fed64cce